backend/requirements.txt
fastapi
uvicorn
google-generativeai
pygetwindow
python-dotenv
sqlalchemy
pydantic
